Make what nohost does and doesn't do a bit clearer up-front
The demo was very cool, and because it said "web server in your browser" I thought it was doing something it wasn't--e.g. registering the browser somehow as a local data source.  After a bit of digging I figured out what it actually did.

Which is still neat, but I think it could be explained more clearly.  This is just a suggestion of the direction of edits to make it so that it would have brought me up to speed quicker, I am not pushing it verbatim.  It really isn't "serving" anything and I think that's an important point...which may have been obvious to some people, but I learned that via trying to use it as a way to make XMLHttpRequests from blocking workers to GUI threads and finding it wasn't the loophole I hoped it might be... :-/